+++ selfdrive/car/card.py
@@ -222,6 +222,7 @@ class Car:
     CS.activateCruise = self.v_cruise_helper._activate_cruise
     CS.latEnabled = self.v_cruise_helper._lat_enabled
     CS.useLaneLineSpeed = self.v_cruise_helper.useLaneLineSpeedApply
+    CS.laneModeSwitchState = self.v_cruise_helper.lane_mode_switch_state
     CS.carrotCruise = 1 if self.v_cruise_helper.carrot_cruise_active else 0
 
     self.CI.CS.softHoldActive = CS.softHoldActive